requirements.txt
fastapi~=0.111.0
uvicorn
motor~=3.4.0
pydantic~=2.7.4

pymongo~=4.7.3
python-dotenv~=1.0.1